@@ -1293,7 +1293,7 @@ namespace Microsoft.Data.SqlClient
                     // not present as the rpcName, as is the case with non-_batchRPCMode. So
                     // input parameters start at parameters[1]. parameters[0] is the actual
                     // T-SQL Statement. rpcName is sp_executesql.
-                    if (_RPCList[i].systemParams.Length > 1)
+                    if (_RPCList[i].systemParams != null && _RPCList[i].systemParams.Length > 1)
                     {
                         _RPCList[i].needsFetchParameterEncryptionMetadata = true;
 
